Update value type as Integer for max_io

+++ ros/api/v0/hosts.py
@@ -51,7 +51,7 @@ class HostsApi(Resource):
     performance_utilization_fields = {
         'cpu': fields.Integer,
         'memory': fields.Integer,
-        'max_io': fields.Float,
+        'max_io': fields.Integer,
         'io_all': fields.Raw
     }
     hosts_fields = {
@@ -211,7 +211,7 @@ class HostDetailsApi(Resource):
     performance_utilization_fields = {
         'cpu': fields.Integer,
         'memory': fields.Integer,
-        'max_io': fields.Float,
+        'max_io': fields.Integer,
         'io_all': fields.Raw
     }
     profile_fields = {
@@ -273,7 +273,7 @@ class HostHistoryApi(Resource):
         'cpu': fields.Integer,
         'memory': fields.Integer,
         'io_all': fields.Raw,
-        'max_io': fields.Float,
+        'max_io': fields.Integer,
         'report_date': fields.String
     }
     meta_fields = {
